# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/tax_transactions:
    get:
      tags:
        - TaxTransactions
      summary: List TaxTransactions
      description: Get a paginated list of TaxTransactions with optional filtering and sorting
      operationId: listTaxTransactions
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxTransactionList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - TaxTransactions
      summary: Create TaxTransaction
      description: Create a new TaxTransaction
      operationId: createTaxTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaxTransactionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxTransaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/withholding_categories:
    get:
      tags:
        - WithholdingCategories
      summary: List WithholdingCategories
      description: Get a paginated list of WithholdingCategories with optional filtering and sorting
      operationId: listWithholdingCategories
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithholdingCategoryList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - WithholdingCategories
      summary: Create WithholdingCategory
      description: Create a new WithholdingCategory
      operationId: createWithholdingCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWithholdingCategoryRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithholdingCategory'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    TaxTransaction:
      type: object
      required:
        - id
        - invoice_ref
        - invoice_kind
        - posting_date
        - taxable_base
        - output_total
        - input_total
        - withholding_total
        - currency
        - source
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        invoice_ref:
          type: string
          format: uuid
        invoice_kind:
          type: string
        posting_date:
          type: string
          format: date
        taxable_base:
          type: string
        output_total:
          type: string
        input_total:
          type: string
        withholding_total:
          type: string
        currency:
          type: string
          maxLength: 3
        source:
          type: string
        efaktur_document_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        metadata:
          type: object
...
    CreateTaxTransactionRequest:
      type: object
      required:
        - invoice_ref
        - invoice_kind
        - posting_date
        - taxable_base
        - output_total
        - input_total
        - withholding_total
        - currency
        - source
        - status
        - metadata
      properties:
        invoice_ref:
          type: string
          format: uuid
        invoice_kind:
          type: string
        posting_date:
          type: string
          format: date
        taxable_base:
          type: string
        output_total:
          type: string
        input_total:
          type: string
        withholding_total:
          type: string
        currency:
          type: string
          maxLength: 3
        source:
          type: string
        efaktur_document_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        metadata:
          type: object
...
    TaxTransactionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TaxTransaction'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    WithholdingCategory:
      type: object
      required:
        - id
        - code
        - name
        - rate
        - threshold_amount
        - effective_from
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 40
        name:
          type: string
          maxLength: 150
        rate:
          type: string
        threshold_amount:
          type: string
        account_id:
          nullable: true
          type: string
          format: uuid
        effective_from:
          type: string
          format: date
        effective_to:
          nullable: true
          type: string
          format: date
        status:
          type: string
        metadata:
          type: object
...
    CreateWithholdingCategoryRequest:
      type: object
      required:
        - code
        - name
        - rate
        - threshold_amount
        - effective_from
        - status
        - metadata
      properties:
        code:
          type: string
          maxLength: 40
        name:
          type: string
          maxLength: 150
        rate:
          type: string
        threshold_amount:
          type: string
        account_id:
          nullable: true
          type: string
          format: uuid
        effective_from:
          type: string
          format: date
        effective_to:
          nullable: true
          type: string
          format: date
        status:
          type: string
        metadata:
          type: object
...
    WithholdingCategoryList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WithholdingCategory'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []